chore: cleaned bdd step definitions, deletes unused methods, and renamed all methods
