+++ lsadf_api/src/test/java/com/lsadf/application/bdd/step_definition/given/BddGivenStepDefinitions.java
@@ -13,19 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.admin.application.bdd.given;
+package com.lsadf.application.bdd.step_definition.given;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.lsadf.admin.application.bdd.BddLoader;
-import com.lsadf.admin.application.bdd.CacheEntryType;
+import com.lsadf.application.bdd.BddLoader;
+import com.lsadf.application.bdd.CacheEntryType;
 import com.lsadf.core.bdd.BddFieldConstants;
 import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
+import com.lsadf.core.domain.game.save.metadata.GameMetadata;
 import com.lsadf.core.domain.game.save.stage.Stage;
 import com.lsadf.core.infra.exception.http.NotFoundException;
-import com.lsadf.core.infra.persistence.table.game.inventory.ItemEntity;
 import com.lsadf.core.infra.persistence.table.game.save.characteristics.CharacteristicsEntity;
 import com.lsadf.core.infra.persistence.table.game.save.currency.CurrencyEntity;
 import com.lsadf.core.infra.persistence.table.game.save.metadata.GameMetadataEntity;
@@ -33,9 +33,9 @@ import com.lsadf.core.infra.persistence.table.game.save.stage.StageEntity;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
 import java.time.Clock;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
@@ -44,89 +44,23 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j(topic = "[GIVEN STEP DEFINITIONS]")
 public class BddGivenStepDefinitions extends BddLoader {
 
-  @Given("^the following items to the inventory of the game save with id (.*)$")
-  @Transactional
-  public void given_the_following_items(String gameSaveId, DataTable dataTable) {
-    List<Map<String, String>> rows = dataTable.asMaps(String.class, String.class);
-
-    UUID uuid = UUID.fromString(gameSaveId);
-    var exists = gameMetadataRepository.existsById(uuid);
-    if (!exists) {
-      throw new NotFoundException("Game save with id: " + gameSaveId + " not found.");
-    }
-
-    log.info("Creating items...");
-    rows.forEach(
-        row -> {
-          ItemEntity itemEntity = BddUtils.mapToItemEntity(row);
-          itemEntity.setGameSaveId(uuid);
-          ItemEntity newItemEntity;
-          if (itemEntity.getId() != null) {
-            newItemEntity =
-                itemRepository.createNewItemEntity(
-                    itemEntity.getId(),
-                    itemEntity.getGameSaveId(),
-                    itemEntity.getClientId(),
-                    itemEntity.getBlueprintId(),
-                    itemEntity.getItemType(),
-                    itemEntity.getItemRarity(),
-                    itemEntity.getIsEquipped(),
-                    itemEntity.getLevel(),
-                    itemEntity.getMainStatistic(),
-                    itemEntity.getMainBaseValue());
-          } else {
-            newItemEntity =
-                itemRepository.createNewItemEntity(
-                    itemEntity.getGameSaveId(),
-                    itemEntity.getClientId(),
-                    itemEntity.getBlueprintId(),
-                    itemEntity.getItemType(),
-                    itemEntity.getItemRarity(),
-                    itemEntity.getIsEquipped(),
-                    itemEntity.getLevel(),
-                    itemEntity.getMainStatistic(),
-                    itemEntity.getMainBaseValue());
-          }
-
-          var additionalStats = BddUtils.mapToAdditionalItemStatEntity(row, newItemEntity.getId());
-          additionalStats.forEach(
-              additionalItemStatEntity -> {
-                additionalItemStatsRepository.createNewAdditionalItemStatEntity(
-                    additionalItemStatEntity.getItemId(),
-                    additionalItemStatEntity.getStatistic(),
-                    additionalItemStatEntity.getBaseValue());
-              });
-          log.info("Item created: {}", newItemEntity);
-        });
-
-    log.info("Items created");
-  }
-
   @Given("^the BDD engine is ready$")
-  public void given_the_bdd_engine_is_ready() {
+  public void givenBddEngineIsReady() {
     BddUtils.initTestRestTemplate(testRestTemplate);
 
     log.info("BDD engine is ready. Using port: {}", this.serverPort);
   }
 
   @Given("^the time clock set to the present$")
-  public void given_the_time_clock_set_to_the_present() {
+  public void givenTimeClockSetToPresent() {
     log.info("Setting time clock to the present...");
     this.clockService.setClock(Clock.systemDefaultZone());
     log.info("Time clock set to the present");
   }
 
-  @Given("^the time clock set to the following value (.*)$")
-  public void given_the_time_clock_set_to_the_following_value(String time) {
-    log.info("Setting time clock to the following value: {}", time);
-    Instant instant = Instant.parse(time);
-    ZoneId zoneId = clockService.getClock().getZone();
-    clockService.setClock(Clock.fixed(instant, zoneId));
-    log.info("Time clock set to the following value: {}", time);
-  }
 
   @Given("^the cache is enabled$")
-  public void given_the_cache_is_enabled() {
+  public void givenCacheIsEnabled() {
     log.info("Checking cache status...");
     boolean cacheEnabled = redisCacheService.isEnabled();
     if (!cacheEnabled) {
@@ -140,7 +74,7 @@ public class BddGivenStepDefinitions extends BddLoader {
   }
 
   @Given("^the cache is disabled$")
-  public void given_the_cache_is_disabled() {
+  public void givenCacheIsDisabled() {
     log.info("Checking cache status...");
     boolean cacheEnabled = redisCacheService.isEnabled();
     if (cacheEnabled) {
@@ -155,7 +89,7 @@ public class BddGivenStepDefinitions extends BddLoader {
 
   @Given("^a clean database$")
   @Transactional
-  public void given_i_have_a_clean_database() throws NotFoundException {
+  public void givenCleanDatabase() throws NotFoundException {
     log.info("Cleaning database repositories...");
 
     this.gameMetadataRepository.deleteAllGameSaveEntities();
@@ -172,11 +106,8 @@ public class BddGivenStepDefinitions extends BddLoader {
     localCacheService.clearCaches();
 
     assertThat(characteristicsCache.getAll()).isEmpty();
-    assertThat(characteristicsCache.getAllHisto()).isEmpty();
     assertThat(currencyCache.getAll()).isEmpty();
-    assertThat(currencyCache.getAllHisto()).isEmpty();
     assertThat(stageCache.getAll()).isEmpty();
-    assertThat(stageCache.getAllHisto()).isEmpty();
     assertThat(gameMetadataCache.getAll()).isEmpty();
 
     log.info("Database repositories + caches cleaned");
@@ -184,47 +115,48 @@ public class BddGivenStepDefinitions extends BddLoader {
   }
 
   @Given("^the following game saves$")
-  public void given_i_have_the_following_game_saves(DataTable dataTable) {
+  @Transactional
+  public void givenFollowingGameSaves(DataTable dataTable) {
     List<Map<String, String>> rows = dataTable.asMaps(String.class, String.class);
     log.info("Creating game saves...");
 
     rows.forEach(
         row -> {
           GameMetadataEntity gameMetadataEntity = BddUtils.mapToGameSaveEntity(row);
+          GameMetadataEntity newEntity =
+              gameMetadataEntity.getId() != null
+                  ? gameMetadataRepository.createNewGameSaveEntity(
+                      gameMetadataEntity.getId(),
+                      gameMetadataEntity.getUserEmail(),
+                      gameMetadataEntity.getNickname())
+                  : gameMetadataRepository.createNewGameSaveEntity(
+                      gameMetadataEntity.getUserEmail(), gameMetadataEntity.getNickname());
           CurrencyEntity currencyEntity = BddUtils.mapToCurrencyEntity(row);
-          CharacteristicsEntity characteristicsEntity = BddUtils.mapToCharacteristicsEntity(row);
-          StageEntity stageEntity = BddUtils.mapToStageEntity(row);
-
-          GameMetadataEntity newGameMetadataEntity =
-              gameMetadataRepository.createNewGameSaveEntity(
-                  gameMetadataEntity.getId(),
-                  gameMetadataEntity.getUserEmail(),
-                  gameMetadataEntity.getNickname());
           currencyRepository.createNewCurrencyEntity(
-              newGameMetadataEntity.getId(),
+              newEntity.getId(),
               currencyEntity.getGoldAmount(),
               currencyEntity.getDiamondAmount(),
               currencyEntity.getEmeraldAmount(),
               currencyEntity.getAmethystAmount());
+          StageEntity stageEntity = BddUtils.mapToStageEntity(row);
           stageRepository.createNewStageEntity(
-              newGameMetadataEntity.getId(),
-              stageEntity.getCurrentStage(),
-              stageEntity.getMaxStage());
+              newEntity.getId(), stageEntity.getCurrentStage(), stageEntity.getMaxStage());
+          CharacteristicsEntity characteristicsEntity = BddUtils.mapToCharacteristicsEntity(row);
           characteristicsRepository.createNewCharacteristicsEntity(
-              newGameMetadataEntity.getId(),
+              newEntity.getId(),
               characteristicsEntity.getAttack(),
               characteristicsEntity.getCritChance(),
               characteristicsEntity.getCritDamage(),
               characteristicsEntity.getHealth(),
               characteristicsEntity.getResistance());
-          log.info("Game save created: {}", newGameMetadataEntity);
+          log.info("Game save created: {}", newEntity);
         });
 
     log.info("Game saves created");
   }
 
   @Given("^the following (.*) entries in cache$")
-  public void given_the_following_cache_entries_in_cache(String cacheType, DataTable dataTable) {
+  public void givenFollowingCacheEntries(String cacheType, DataTable dataTable) {
     List<Map<String, String>> rows = dataTable.asMaps(String.class, String.class);
     log.info("Creating {} entries in cache...", cacheType);
     CacheEntryType cacheEntryType = CacheEntryType.fromString(cacheType);
@@ -253,6 +185,17 @@ public class BddGivenStepDefinitions extends BddLoader {
                 String gameSaveId = row.get(BddFieldConstants.StageCacheEntry.GAME_SAVE_ID);
                 Stage stage = BddUtils.mapToStage(row);
                 stageCache.set(gameSaveId, stage);
+                count.getAndIncrement();
+              });
+      case GAME_METADATA ->
+          rows.forEach(
+              row -> {
+                String gameSaveId =
+                    row.get(BddFieldConstants.GameSaveOwnershipCacheEntry.GAME_SAVE_ID);
+                String userId = row.get(BddFieldConstants.GameSaveOwnershipCacheEntry.USER_EMAIL);
+                GameMetadata gameMetadata =
+                    new GameMetadata(UUID.fromString(gameSaveId), userId, null, null, null);
+                gameMetadataCache.set(gameSaveId, gameMetadata);
                 count.getAndIncrement();
               });
       default -> throw new IllegalArgumentException("Unknown cache type: " + cacheType);
